rename: move api interface to a new package
