+++ app/src/main/java/com/hapataka/questwalk/data/retrofit/WeatherService.kt
@@ -1,4 +1,4 @@
-package com.hapataka.questwalk.data.datasource.remote
+package com.hapataka.questwalk.data.retrofit
 
 import com.hapataka.questwalk.data.dto.WeatherResponse
 import retrofit2.http.GET
